Tweak regular expression to match @-replies.
This is tricky. If we use the regular expression that finds a `<p>` tag,
we'll miss situations in which users reply to prior comments but do so
after the beginning of their comment.

On the other hand, if we don't use the `<p>` in the regular expression,
we get a situation where every comment on the "improvements" pages are
matched because they contain a `<time>` element in which there is an `@`
symbol.

An easier solution is to be stricter about the @-reply matching, and by
simply removing the optional whitespace characters between the `@` and
the nickname, the code will work across the board. So, let's do that.